+++ Navferty.Common/Extensions/ControlsExtensions.cs
@@ -69,152 +69,6 @@ namespace Navferty.Common
 			tmrDelay.Start();
 		}
 
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		internal static StringAlignment GetAlignment(this ContentAlignment ca)
-			=> ca switch
-			{
-				var ca2
-				when ca2 == ContentAlignment.TopLeft || ca2 == ContentAlignment.MiddleLeft || ca2 == ContentAlignment.BottomLeft
-				=> StringAlignment.Near,
-
-				var ca2
-				when ca2 == ContentAlignment.TopCenter || ca2 == ContentAlignment.MiddleCenter || ca2 == ContentAlignment.BottomCenter
-				=> StringAlignment.Center,
-
-				var ca2
-				when ca2 == ContentAlignment.TopRight || ca2 == ContentAlignment.MiddleRight || ca2 == ContentAlignment.BottomRight
-					=> StringAlignment.Far,
-
-				_ => StringAlignment.Center,
-			};
-
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		internal static StringAlignment GetLineAlignment(this ContentAlignment ca)
-			=> ca switch
-			{
-				var ca2
-				when ca2 == ContentAlignment.TopLeft || ca2 == ContentAlignment.TopCenter || ca2 == ContentAlignment.TopRight
-				=> StringAlignment.Near,
-
-				var ca2
-				when ca2 == ContentAlignment.MiddleLeft || ca2 == ContentAlignment.MiddleCenter || ca2 == ContentAlignment.MiddleRight
-				=> StringAlignment.Center,
-
-				var ca2
-				when ca2 == ContentAlignment.BottomLeft || ca2 == ContentAlignment.BottomCenter || ca2 == ContentAlignment.BottomRight
-					=> StringAlignment.Far,
-
-				_ => StringAlignment.Center,
-			};
-
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		internal static StringFormat ToStringFormat(this ContentAlignment ca)
-		{
-			var sf = new StringFormat()
-			{
-				Alignment = ca.GetAlignment(),
-				LineAlignment = ca.GetLineAlignment()
-			};
-			return sf;
-		}
-
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		internal static void DrawTextEx(
-			this Graphics g,
-			string text,
-			Font font,
-			Color textcolor,
-			Rectangle rect,
-			ContentAlignment textAlign,
-			StringFormatFlags? additionalFF = null)
-		{
-			using (Brush brText = new SolidBrush(textcolor))
-				g.DrawTextEx(text, font, brText, rect, textAlign, additionalFF);
-		}
-
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		internal static void DrawTextEx(
-			this Graphics g,
-			string text,
-			Font font,
-			Brush textbrush,
-			Rectangle rect,
-			ContentAlignment textAlign,
-			StringFormatFlags? additionalFF = null)
-		{
-			using (var sf = textAlign.ToStringFormat())
-			{
-				if (additionalFF != null && additionalFF.HasValue) sf.FormatFlags |= additionalFF.Value;
-				g.DrawString(text, font, textbrush, rect, sf);
-			}
-		}
-
-		/*
-		 
-		internal static GraphicsPath CreateRoundRect(this Graphics g, Rectangle rect)
-		{
-			GraphicsPath path = new();
-			path.AddArc(rect.Left, rect.Top, rect.Height, rect.Height, 90, 180);
-			path.AddArc(rect.Right - rect.Height, rect.Top, rect.Height, rect.Height, -90, 180);
-			path.CloseFigure();
-			return path;
-		}
-		*/
-
-		[DebuggerNonUserCode, DebuggerStepThrough, MethodImpl(MethodImplOptions.AggressiveInlining)]
-		internal static void DrawRoundRect(
-			   this Graphics g,
-			   Rectangle rc,
-			   int radius,
-			   Pen? pen = null,
-			   Brush? brush = null)
-		{
-			Rectangle rcCorner = new(rc.X, rc.Y, radius, radius);
-
-			using (System.Drawing.Drawing2D.GraphicsPath path = new())
-			{
-				path.AddArc(rcCorner, 180, 90);
-
-				rcCorner.X = rc.X + rc.Width - radius;
-				path.AddArc(rcCorner, 270, 90);
-
-				rcCorner.Y = rc.Y + rc.Height - radius;
-				path.AddArc(rcCorner, 0, 90);
-
-				rcCorner.X = rc.X;
-				path.AddArc(rcCorner, 90, 90);
-
-				path.CloseFigure();
-
-				if (brush != null) g.FillPath(brush, path);
-				if (pen != null) g.DrawPath(pen, path);
-			}
-		}
-
-
-
-		[DebuggerNonUserCode, DebuggerStepThrough, MethodImpl(MethodImplOptions.AggressiveInlining)]
-		internal static void DeflateByPadding(
-			   this ref Rectangle rect,
-			   Padding pd)
-		{
-			if (null == pd) return;
-
-			if (pd.Left > 0)
-			{
-				rect.Offset(pd.Left, 0);
-				rect.Width -= pd.Left;
-			}
-			if (pd.Right > 0) rect.Width -= pd.Right;
-
-			if (pd.Top > 0)
-			{
-				rect.Offset(0, pd.Top);
-				rect.Height -= pd.Top;
-			}
-			if (pd.Bottom > 0) rect.Height -= pd.Bottom;
-		}
-
 
 		#region AttachDelayedFilter
